refactor: split jsonrpc dispatcher
