+++ pkgs/standards/autoapi/autoapi/v3/transport/jsonrpc/dispatcher.py
@@ -78,116 +78,21 @@ except Exception:  # pragma: no cover
 from ...runtime.errors import ERROR_MESSAGES, http_exc_to_rpc
 from ...config.constants import AUTOAPI_AUTH_CONTEXT_ATTR
 from .models import RPCRequest, RPCResponse
+from .helpers import (
+    _authorize,
+    _err,
+    _model_for,
+    _normalize_deps,
+    _normalize_params,
+    _ok,
+    _select_auth_dep,
+    _user_from_request,
+)
 
 logger = logging.getLogger(__name__)
 
 Json = Mapping[str, Any]
 Batch = Sequence[Mapping[str, Any]]
-
-# --------------------------------------------------------------------------- #
-# Utilities
-# --------------------------------------------------------------------------- #
-
-
-def _ok(result: Any, id_: Any) -> Dict[str, Any]:
-    return {"jsonrpc": "2.0", "result": result, "id": id_}
-
-
-def _err(code: int, msg: str, id_: Any, data: Any | None = None) -> Dict[str, Any]:
-    e: Dict[str, Any] = {
-        "jsonrpc": "2.0",
-        "error": {"code": code, "message": msg},
-        "id": id_,
-    }
-    if data is not None:
-        e["error"]["data"] = data
-    return e
-
-
-def _normalize_params(params: Any) -> Any:
-    if params is None:
-        return {}
-    if isinstance(params, Mapping):
-        return dict(params)
-    if isinstance(params, Sequence) and not isinstance(params, (str, bytes)):
-        return list(params)
-    raise HTTPException(
-        status_code=400, detail="Invalid params: expected object or array"
-    )
-
-
-def _model_for(api: Any, name: str) -> Optional[type]:
-    models: Dict[str, type] = getattr(api, "models", {}) or {}
-    mdl = models.get(name)
-    if mdl is not None:
-        return mdl
-    # Case-insensitive fallback
-    lower = name.lower()
-    for k, v in models.items():
-        if k.lower() == lower:
-            return v
-    return None
-
-
-def _user_from_request(request: Request) -> Any | None:
-    return getattr(request.state, "user", None)
-
-
-def _select_auth_dep(api: Any):
-    """
-    Choose the appropriate auth dependency based on API flags.
-    Order:
-      1) optional if provided,
-      2) required when allow_anon == False and _authn exists,
-      3) otherwise _authn if present,
-      4) else None.
-    """
-    if getattr(api, "_optional_authn_dep", None):
-        return api._optional_authn_dep
-    if getattr(api, "_allow_anon", True) is False and getattr(api, "_authn", None):
-        return api._authn
-    if getattr(api, "_authn", None):
-        return api._authn
-    return None
-
-
-def _normalize_deps(deps: Optional[Sequence[Any]]) -> list:
-    """Accept either Depends(...) objects or plain callables."""
-    out = []
-    for d in deps or ():
-        try:
-            is_dep_obj = hasattr(d, "dependency")
-        except Exception:
-            is_dep_obj = False
-        out.append(d if is_dep_obj else Depends(d))
-    return out
-
-
-def _authorize(
-    api: Any,
-    request: Request,
-    model: type,
-    alias: str,
-    payload: Mapping[str, Any],
-    user: Any | None,
-):
-    """
-    Call an authorize gate if present. Prefer api._authorize; else model.__autoapi_authorize__.
-    Falsy returns or exceptions (non-HTTPException) become 403.
-    """
-    fn = getattr(api, "_authorize", None) or getattr(
-        model, "__autoapi_authorize__", None
-    )
-    if not fn:
-        return
-    try:
-        rv = fn(request=request, model=model, alias=alias, payload=payload, user=user)
-        if rv is False:
-            raise HTTPException(status_code=403, detail="Forbidden")
-    except HTTPException:
-        raise
-    except Exception:
-        raise HTTPException(status_code=403, detail="Forbidden")
 
 
 async def _dispatch_one(
